lab 24/09/24 take 2

+++ MeaningfulFunctions2/MeaningfulFunctions2/MeaningfulFunctions2/Calculations.h
@@ -1,6 +1,6 @@
 #pragma once
 
 int product(int i, int j);
-double quotient(int d, int i);
+double quotient(double d, int i);
 int add(int i, int j);
 int subtract(int i, int j);
