+++ MeaningfulFunctions2/MeaningfulFunctions2/MeaningfulFunctions2/Source.cpp
@@ -2,7 +2,7 @@
 
 int product(int i, int j) { return i * j; }
 
-double quotient(int d, int i) { return d / i; }
+double quotient(double d, int i) { return d / i; }
 
 int add(int i, int j) { return i + j; }
